220329 망할 페이징url 강제형변환

+++ src/main/java/com/kh146/domain/board/Bbs.java
@@ -13,7 +13,7 @@ public class Bbs {
   private String authorId;
   private String nickname;
   private String bcontent;
-  private Long hit;
+  private int hit;
   private LocalDate cdate;
   private LocalDate udate;
